refactor(posthog-ai): share tool output parsing
